+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - CLI flag `--env-file` added to the `blobl` command. (@mihaitodor)
 - Go API: New `ComponentLinter` APIs for linting general component configs. (@Jeffail)
 - New `bitwise_and`, `bitwise_or`, and `bitwise_xor` bloblang methods. (@eadwright)
+- Field `open_message_mapping` added to the `socket` input. (@eadwright)
 
 ### Fixed